RavenDB-22873 - unify the logic for index name verification
