+++ src/Raven.Server/ServerWide/Commands/Indexes/PutIndexCommand.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.ServerWide;
+using Raven.Server.Documents.Indexes;
 using Raven.Server.Extensions;
 using Raven.Server.Rachis;
 using Sparrow.Json.Parsing;
@@ -38,15 +40,11 @@ namespace Raven.Server.ServerWide.Commands.Indexes
 
         public override void UpdateDatabaseRecord(DatabaseRecord record, long etag)
         {
+            var indexValidator = new StaticIndexNameValidator(record);
+            indexValidator.Validate(Definition);
+
             try
             {
-                var indexNames = record.Indexes.Select(x => x.Value.Name).ToHashSet(StringComparer.OrdinalIgnoreCase);
-
-                if (indexNames.Add(Definition.Name) == false && record.Indexes.TryGetValue(Definition.Name, out _) == false)
-                {
-                    throw new InvalidOperationException($"Can not add index: {Definition.Name} because an index with the same name but different casing already exist");
-                }
-
                 record.AddIndex(Definition, Source, CreatedAt, etag, RevisionsToKeep, DefaultDeploymentMode ?? IndexDeploymentMode.Parallel);
 
             }
@@ -74,6 +72,47 @@ namespace Raven.Server.ServerWide.Commands.Indexes
             }
 
             return msg;
+        }
+
+        public class StaticIndexNameValidator
+        {
+            private readonly DatabaseRecord _record;
+            private HashSet<string> _indexNames;
+            private HashSet<string> _safeFileSystemIndexNames;
+
+            public StaticIndexNameValidator(DatabaseRecord record)
+            {
+                _record = record;
+            }
+
+            public void Validate(IndexDefinition definition)
+            {
+                if (_record.Indexes.TryGetValue(definition.Name, out _))
+                    return;
+
+                // this is not an update to an existing index. we'll check for:
+                // - directory name collisions
+                // - index name case sensitivity
+
+                _indexNames ??= _record.Indexes.Select(x => x.Value.Name).ToHashSet(StringComparer.OrdinalIgnoreCase);
+                _safeFileSystemIndexNames ??= _record.Indexes.Select(x => IndexDefinitionBaseServerSide.GetIndexNameSafeForFileSystem(x.Value.Name)).ToHashSet(StringComparer.OrdinalIgnoreCase);
+
+                if (_indexNames.Add(definition.Name) == false)
+                {
+                    throw new RachisApplyException($"Can not add index: {definition.Name} because an index with the same name but different casing already exist");
+                }
+
+                var safeFileSystemIndexName = IndexDefinitionBaseServerSide.GetIndexNameSafeForFileSystem(definition.Name);
+                if (_safeFileSystemIndexNames.Add(safeFileSystemIndexName) == false)
+                {
+                    var existingIndexName = _indexNames.FirstOrDefault(x =>
+                        x.Equals(definition.Name, StringComparison.OrdinalIgnoreCase) == false &&
+                        safeFileSystemIndexName.Equals(IndexDefinitionBaseServerSide.GetIndexNameSafeForFileSystem(x), StringComparison.OrdinalIgnoreCase));
+
+                    throw new RachisApplyException(
+                        $"Could not create index '{definition.Name}' because it would result in directory name collision with '{existingIndexName}' index");
+                }
+            }
         }
     }
 }
